# api/docs/swagger.yaml
definitions:
...
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
  users_service.AdminResponse:
    properties:
      file:
        type: string
      first_name:
        type: string
      last_message:
        type: string
      last_name:
        type: string
      message_count:
        type: integer
      user_id:
        type: string
    type: object
...
  users_service.GetMessageAdminResponse:
    properties:
      admin_message:
        items:
          $ref: '#/definitions/users_service.AdminResponse'
        type: array
      count:
        type: integer
    type: object
  users_service.GetMessageUserResponse:
    properties:
      count:
        type: integer
      messages:
        items:
          $ref: '#/definitions/users_service.Message'
        type: array
    type: object
  users_service.Message:
    properties:
      created_at:
        type: string
      file:
        type: string
      id:
        type: string
      read:
        type: string
      status:
        type: string
      text:
        type: string
      user_id:
        type: string
    type: object
...
  users_service.UpdateMessageRequest:
    properties:
      id:
        type: string
      read:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /admin/message:
    post:
      consumes:
      - multipart/form-data
      description: Create Message
      operationId: create_admin_message
      parameters:
      - description: Upload file
        in: formData
        name: file
        type: file
      - description: TEXT
        in: formData
        name: text
        required: true
        type: string
      - description: UserId
        in: formData
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetCoinBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Message
      tags:
      - Admin
  /admin/message/:
    get:
      consumes:
      - application/json
      description: Get Admin Message  By ID
      operationId: get_admin_all_message_by_id
      produces:
      - application/json
      responses:
        "200":
          description: UserMessageBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/users_service.GetMessageAdminResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Admin Message  By ID
      tags:
      - Admin Message
  /admin/message/{id}:
    get:
      consumes:
      - application/json
      description: Get Admin Message  By ID
      operationId: get_admin_message_by_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: UserMessageBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/users_service.GetMessageUserResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Admin Message  By ID
      tags:
      - Admin Message
  /admin/message/user/{id}:
    get:
      consumes:
      - application/json
      description: Get Admin Message  By ID
      operationId: get_message_admin_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: UserMessageBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/users_service.GetMessageUserResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Admin Message  By ID
      tags:
      - Admin Message
...
  /message/{id}:
    put:
      consumes:
      - application/json
      description: Update Message
      operationId: update_message
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateMessage
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/users_service.UpdateMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Message data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Message
      tags:
      - Message
...
  /user/message:
    post:
      consumes:
      - multipart/form-data
      description: Create Message
      operationId: create_user_message
      parameters:
      - description: Upload file
        in: formData
        name: file
        type: file
      - description: TEXT
        in: formData
        name: text
        required: true
        type: string
      - description: UserId
        in: formData
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetCoinBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Message
      tags:
      - User
  /user/message/{id}:
    get:
      consumes:
      - application/json
      description: Get User Message  By ID
      operationId: get_user_message_by_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: UserMessageBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/users_service.GetMessageUserResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get User Message  By ID
      tags:
      - User Message
...
swagger: "2.0"